backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
bcrypt==4.1.2
python-dotenv==1.0.1
email-validator==2.1.1
joblib==1.3.2
waitress==3.0.0
openai==1.12.0
